+++ src/textual/_border.py
@@ -237,6 +237,8 @@ BoxSegments: TypeAlias = Tuple[
 
 Borders: TypeAlias = Tuple[EdgeStyle, EdgeStyle, EdgeStyle, EdgeStyle]
 
+REVERSE_STYLE = Style(reverse=True)
+
 
 @lru_cache(maxsize=1024)
 def get_box(
@@ -249,9 +251,9 @@ def get_box(
 
     Args:
         name: Name of the box type.
-        inner_style: The inner style (widget background)
-        outer_style: The outer style (parent background)
-        style: Widget style
+        inner_style: The inner style (widget background).
+        outer_style: The outer style (parent background).
+        style: Widget style.
 
     Returns:
         A tuple of 3 Segment triplets.
@@ -271,11 +273,12 @@ def get_box(
 
     inner = inner_style + style
     outer = outer_style + style
+
     styles = (
         inner,
         outer,
-        Style.from_color(outer.bgcolor, inner.color),
-        Style.from_color(inner.bgcolor, outer.color),
+        Style.from_color(inner.color, outer.bgcolor) + REVERSE_STYLE,
+        Style.from_color(outer.color, inner.bgcolor) + REVERSE_STYLE,
     )
 
     return (
@@ -363,9 +366,9 @@ def render_border_label(
     elif label_style_location == 1:
         base_style = outer
     elif label_style_location == 2:
-        base_style = Style.from_color(outer.bgcolor, inner.color)
+        base_style = Style.from_color(inner.color, outer.bgcolor) + REVERSE_STYLE
     elif label_style_location == 3:
-        base_style = Style.from_color(inner.bgcolor, outer.color)
+        base_style = Style.from_color(outer.color, inner.bgcolor) + REVERSE_STYLE
     else:
         assert False