+++ src/textual/widgets/_footer.py
@@ -69,7 +69,7 @@ class FooterKey(Widget):
             .footer-key--description {
                 padding: 0 0 0 1;
             }
-        }
+        }        
     }
     """
 
@@ -147,7 +147,31 @@ class Footer(ScrollableContainer, can_focus=False, can_focus_children=False):
         FooterKey.-command-palette  {
             dock: right;                        
             padding-right: 1;
-            border-left: vkey $foreground 20%;                
+            border-left: vkey $foreground 20%;                            
+        }
+
+        &.-ansi-colors {           
+            background: ansi_default;            
+            .footer-key--key {
+                background: ansi_default;
+                color: ansi_magenta;                
+            }
+            .footer-key--description {
+                background: ansi_default;
+                color: ansi_default;                
+            }
+            FooterKey:hover {
+                text-style: underline;
+                background: ansi_default;
+                color: ansi_default;
+                .footer-key--key {
+                    background: ansi_default;
+                }
+            }        
+            FooterKey.-command-palette {
+                background: ansi_default;
+                border-left: vkey ansi_black;
+            }
         }
     }
     """